requirements.txt
monero-python>=0.4.3
